add findByEmail upd security config
