+++ src/main/java/com/simbirsoft/projectManager/rest/controller/UserController.java
@@ -1,4 +1,4 @@
-package com.simbirsoft.projectManager.rest;
+package com.simbirsoft.projectManager.rest.controller;
 
 import com.simbirsoft.projectManager.rest.dto.request.UserRegisterRequest;
 import com.simbirsoft.projectManager.rest.dto.response.UserResponse;
@@ -20,13 +20,6 @@ public class UserController {
 
     public UserController(UserService userService) {
         this.userService = userService;
-    }
-
-    @Operation(summary = "Авторизация пользователя")
-    @PostMapping("/login")
-    public ResponseEntity<Boolean> login() {
-        // TODO: complete login through Spring Security
-        return ResponseEntity.ok(false);
     }
 
     @Operation(summary = "Регистрация пользователя")
