+++ src/main/java/com/simbirsoft/projectManager/security/service/UserAuthService.java
@@ -1,4 +1,4 @@
-package com.simbirsoft.projectManager.service.impl;
+package com.simbirsoft.projectManager.security.service;
 
 import com.simbirsoft.projectManager.entity.User;
 import com.simbirsoft.projectManager.exception.NotFoundException;
@@ -23,12 +23,13 @@ public class UserAuthService implements UserDetailsService {
     }
 
     @Override
-    public UserDetails loadUserByUsername(String username) {
-        Optional<User> user = repository.findByFullName(username);
-        if (user.isPresent()) {
+    public UserDetails loadUserByUsername(String email) {
+        Optional<User> user = repository.findByEmail(email);
+        if (user.isPresent() ) {
             return new org.springframework.security.core.userdetails.User(user.get().getFullName(),
                     user.get().getPassword(), Collections.singletonList(new SimpleGrantedAuthority("USER")));
         }
-        throw new NotFoundException(User.class, "fullName", username);
+        throw new NotFoundException(User.class, "fullName", email);
     }
+    //if we will keep the role in user table new SimpleGrantedAuthority(user.get().getRole()))
 }
