+++ src/main/java/com/simbirsoft/projectManager/service/UserService.java
@@ -4,10 +4,15 @@ import com.simbirsoft.projectManager.rest.dto.request.UserRegisterRequest;
 import com.simbirsoft.projectManager.rest.dto.response.UserResponse;
 
 import java.util.UUID;
+import org.springframework.transaction.annotation.Transactional;
 
 public interface UserService {
 
     UserResponse getUserById(String id);
+
+
+    @Transactional(readOnly = true)
+    UserResponse findByEmail(String email);
 
     UUID registerUser(UserRegisterRequest request);
