+++ src/main/resources/application.properties
@@ -8,7 +8,7 @@ spring.datasource.hikari.maximumPoolSize=5
 ## PostgreSQL
 spring.datasource.url=jdbc:postgresql://localhost:5432/management_db
 spring.datasource.username=postgres
-spring.datasource.password=zzzzzzzz
+spring.datasource.password=12345
 
 logging.level.root=INFO
 logging.level.org.springframework.security=DEBUG
