+++ src/main/resources/db/changelog/db.changelog-master.yaml
@@ -1,5 +1,4 @@
 databaseChangeLog:
-
   - include: { relativeToChangelogFile: true, file: migrations/changelog-1.0.yaml }
   - include: { relativeToChangelogFile: true, file: migrations/changelog-2.0.yaml }
   - include: { relativeToChangelogFile: true, file: migrations/changelog-2.1.yaml }
