+++ src/main/resources/liquibase.properties
@@ -1,6 +1,6 @@
 url=jdbc:postgresql://localhost:5432/management_db
 username=postgres
-password=zzzzzzzz
+password=12345
 driver=org.postgresql.Driver
-diffChangeLogFile=src/main/resources/db/changelog/migrations/ChangeLog-2.1.yaml
+diffChangeLogFile=src/main/resources/db/changelog/migrations/changelog-4.0.yaml
 diffTypes=tables, views, columns, indexes, foreignkeys, primarykeys, uniqueconstraints, data
